purpose_test/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.5.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
flatbuffers==2.0
frozenlist==1.3.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.44.0
gunicorn==20.1.0
h2==3.2.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
idna==2.10
isort==5.8.0
itsdangerous==2.1.1
Janome==0.4.2
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
libclang==13.0.0
line-bot-sdk==2.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.7.4
PyPDF2==1.26.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.12.1
yarl==1.7.2
